+++ src/test/org/apache/commons/math/distribution/NormalDistributionTest.java
@@ -149,8 +149,8 @@ public class NormalDistributionTest extends ContinuousDistributionAbstractTest  
         distribution.setMean(0);
         distribution.setStandardDeviation(1);
         for (int i = 0; i < 100; i+=5) { // make sure no convergence exception
-            double lowerTail = distribution.cumulativeProbability((double)-i);
-            double upperTail = distribution.cumulativeProbability((double) i);
+            double lowerTail = distribution.cumulativeProbability(-i);
+            double upperTail = distribution.cumulativeProbability(i);
             if (i < 10) { // make sure not top-coded
                 assertTrue(lowerTail > 0.0d);
                 assertTrue(upperTail < 1.0d);
